+++ sanity/schemas/modules/creative-module.ts
@@ -1,115 +1,28 @@
 import { defineArrayMember, defineField, defineType } from 'sanity'
-import {
-	VscExtensions,
-	VscInspect,
-	VscSymbolKeyword,
-	VscVerified,
-} from 'react-icons/vsc'
-import { IoIosImage } from 'react-icons/io'
+import creativeCtas from './creative/creativeCtas'
+import creativeIcon from './creative/creativeIcon'
+import creativeImage from './creative/creativeImage'
+import creativeRichtext from './creative/creativeRichtext'
 import { getBlockText } from '../../src/utils'
 
-const subModules = [
-	defineArrayMember({
-		name: 'ctas',
-		title: 'Call-to-actions',
-		icon: VscInspect,
-		type: 'object',
-		fields: [
-			defineField({
-				name: 'ctas',
-				title: 'Call-to-actions',
-				type: 'array',
-				of: [{ type: 'cta' }],
-			}),
-		],
-		preview: {
-			select: {
-				ctas: 'ctas',
-			},
-			prepare: ({ ctas }) => ({
-				title: `${ctas?.length || 0} CTAs`,
-			}),
-		},
-	}),
-	defineArrayMember({
-		name: 'icon',
-		icon: VscVerified,
-		type: 'object',
-		fields: [
-			defineField({
-				name: 'icon',
-				type: 'image',
-			}),
-			defineField({
-				name: 'height',
-				type: 'number',
-				description: 'px',
-				validation: (Rule) => Rule.min(1).integer(),
-				initialValue: 60,
-			}),
-		],
-		preview: {
-			select: {
-				media: 'icon',
-			},
-		},
-	}),
-	defineArrayMember({
-		name: 'image',
-		icon: IoIosImage,
-		type: 'image',
-		fields: [
-			defineField({
-				name: 'alt',
-				type: 'string',
-			}),
-			defineField({
-				name: 'aspectRatio',
-				type: 'string',
-			}),
-		],
-	}),
-	defineArrayMember({
-		name: 'richtext',
-		icon: VscSymbolKeyword,
-		type: 'object',
-		fields: [
-			defineField({
-				name: 'content',
-				type: 'array',
-				of: [{ type: 'block' }],
-			}),
-		],
-		preview: {
-			select: {
-				content: 'content',
-			},
-			prepare: ({ content }) => ({
-				title: getBlockText(content),
-				subtitle: 'Richtext',
-			}),
-		},
-	}),
-]
+import { VscExtensions } from 'react-icons/vsc'
 
 export default defineType({
 	name: 'creative-module',
 	title: 'Creative module',
 	icon: VscExtensions,
 	type: 'object',
+	groups: [
+		{ title: 'Content', name: 'content', default: true },
+		{ title: 'Options', name: 'options' },
+	],
 	fields: [
 		defineField({
 			name: 'content',
 			type: 'array',
 			of: [{ type: 'block' }],
+			group: 'content',
 		}),
-		defineField({
-			name: 'columns',
-			type: 'number',
-			description: 'Leave empty to use the number of modules as columns',
-			validation: (Rule) => Rule.min(1).integer(),
-		}),
-		// TODO: vertical alignment
 		defineField({
 			name: 'modules',
 			type: 'array',
@@ -118,11 +31,16 @@ export default defineType({
 					title: 'module',
 					type: 'object',
 					fields: [
-						// TODO: column span
 						defineField({
 							name: 'subModules',
 							type: 'array',
-							of: subModules,
+							of: [creativeCtas, creativeIcon, creativeImage, creativeRichtext],
+						}),
+						defineField({
+							name: 'colSpan',
+							title: 'Column span',
+							type: 'number',
+							validation: (Rule) => Rule.min(1).integer(),
 						}),
 					],
 					preview: {
@@ -143,6 +61,37 @@ export default defineType({
 					},
 				}),
 			],
+			group: 'content',
+		}),
+		defineField({
+			name: 'columns',
+			type: 'number',
+			description: 'Leave empty to use the number of modules as columns',
+			validation: (Rule) => Rule.min(1).integer(),
+			group: 'options',
+		}),
+		defineField({
+			name: 'textAlign',
+			type: 'string',
+			options: {
+				list: ['left', 'center', 'right'],
+			},
+			initialValue: 'left',
+			group: 'options',
+		}),
+		defineField({
+			name: 'alignItems',
+			title: 'Vertical alignment',
+			type: 'string',
+			options: {
+				list: [
+					{ title: 'Top', value: 'start' },
+					'center',
+					{ title: 'Bottom', value: 'end' },
+				],
+			},
+			initialValue: 'center',
+			group: 'options',
 		}),
 	],
 	preview: {